--- a/iree/test/microbenchmarks/BUILD
+++ b/iree/test/microbenchmarks/BUILD
@@ -23,11 +23,20 @@ iree_lit_test_suite(
             "linalg_transpose.mlir",
         ],
         include =
-            ["*.mlir"],
+            [
+                "*.mlir"
+            ],
+        exclude =
+            [
+                "linalg_mmt4d.mlir",
+                "mhlo_conv.mlir",
+                "mhlo_dot_general.mlir",
+                "mhlo_dot.mlir",
+                "mhlo_fft_abs.mlir",
+            ],
     ),
     tags = [
         "hostonly",
-        "optonly",  # swiftshader is too slow in dbg
     ],
     tools = [
         "//iree/tools:iree-run-mlir",