--- a/iree/test/microbenchmarks/BUILD
+++ b/iree/test/microbenchmarks/BUILD
@@ -24,13 +24,13 @@ iree_lit_test_suite(
         ],
         include = ["*.mlir"],
         exclude =
-        [
-            "linalg_mmt4d.mlir",
-            "mhlo_conv.mlir",
-            "mhlo_dot_general.mlir",
-            "mhlo_dot.mlir",
-            "mhlo_fft_abs.mlir",
-        ],
+            [
+                "linalg_mmt4d.mlir",
+                "mhlo_conv.mlir",
+                "mhlo_dot_general.mlir",
+                "mhlo_dot.mlir",
+                "mhlo_fft_abs.mlir",
+            ],
     ),
     tags = [
         "hostonly",