--- a/iree/test/microbenchmarks/CMakeLists.txt
+++ b/iree/test/microbenchmarks/CMakeLists.txt
@@ -19,7 +19,6 @@ iree_lit_test_suite(
     iree::tools::iree-run-mlir
   LABELS
     "hostonly"
-    "optonly"
 )
 
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###