This patch adds lit tests and benchmarks for 3-d transpose operations
targeting x86 and AVX2.