@dcaballe 2022-04-06
Some performance numbers:

Transpose Test | Baseline (ms) | 8x8 AVX2 Lowering | Speed-up
-- | -- | -- | --
512x1024 [1, 0] | 0.659 | 0.267 | 2.405
64x96x128 [0, 2, 1] | 0.743 | 0.358 | 2.064
64x96x128 [2, 0, 1] | 0.685 | 0.375 | 1.803
64x96x128 [2, 1, 0] | 0.725 | 0.365 | 1.981
64x96x128 [1, 2, 0] | 0.712 | 0.333 | 2.088
64x96x128 [1, 0, 2] | 0.664 | 0.338 | 1.876